Move UpdateHtmlBuffer
